requirements.txt
mcp>=1.0.0
mysql-connector-python>=8.0.33
python-dotenv>=1.0.0
pyyaml>=6.0
httpx>=0.27.0
asyncio>=3.4.3
pydantic>=2.0.0
uvicorn>=0.29.0
fastapi>=0.111.0
